--- keyboards/trailmix/solecito.c
@@ -1,1 +1,0 @@
-#include "solecito.h"
